backend/functions/process_drive_change/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-vision==3.*
google-cloud-firestore==2.*
google-cloud-bigquery==3.*
